Reorganize project into MVVM structure
