+++ SnapText/App/SnapTextApp.swift
