+++ SnapText/Models/SavedDoc.swift
