+++ SnapText/Services/OCRService.swift
