+++ SnapText/Services/TableDetector.swift
