+++ SnapText/Views/Components/Camera.swift
