+++ SnapText/Views/Components/CropOverlayView.swift
