+++ SnapText/Views/Components/CustomCameraView.swift
