+++ SnapText/Views/Components/CustomCameraViewController.swift
