+++ SnapText/Views/Components/FloatingTextEditingCanvas.swift
