+++ SnapText/Views/Components/ShareSheet.swift
