+++ SnapText/Views/Components/TableEditor.swift
