+++ SnapText/Views/DocsGalleryView.swift
