+++ SnapText/Views/EditableDocView.swift
